Extracted logic into CurrencyLocaleResolver helper, reducing duplicate code and the need for mocking Purchases
