+++ purchases/src/defaults/kotlin/com/revenuecat/purchases/Purchases.kt
@@ -129,20 +129,6 @@ class Purchases internal constructor(
         }
 
     /**
-     * Determines the Locale used for formatting currencies based on the `storefrontCountryCode` and the device's
-     * available locale's. The `storefrontCountryCode` argument will be used instead of the cached
-     * `storefrontCountryCode` if provided. The `locale` argument is used as fallback in case no
-     * `storefrontCountryCode` is available or when there are no matching device locale's.
-     */
-    @InternalRevenueCatAPI
-    fun currencyLocaleForStorefrontCountryCode(
-        storefrontCountryCode: String? = null,
-        locale: Locale = Locale.getDefault(),
-    ): Locale {
-        return purchasesOrchestrator.currencyLocaleForStorefrontCountryCode(storefrontCountryCode, locale)
-    }
-
-    /**
      * Listener that receives callbacks for Customer Center events such as restore initiated,
      * subscription cancellations, and customer feedback submission.
      *
