+++ purchases/src/main/kotlin/com/revenuecat/purchases/models/StoreProduct.kt
@@ -4,6 +4,7 @@ import com.revenuecat.purchases.InternalRevenueCatAPI
 import com.revenuecat.purchases.PresentedOfferingContext
 import com.revenuecat.purchases.ProductType
 import com.revenuecat.purchases.Purchases
+import com.revenuecat.purchases.utils.CurrencyLocaleResolver
 import com.revenuecat.purchases.utils.pricePerDay
 import com.revenuecat.purchases.utils.pricePerMonth
 import com.revenuecat.purchases.utils.pricePerWeek
@@ -211,10 +212,15 @@ interface StoreProduct {
         @OptIn(InternalRevenueCatAPI::class)
         @JvmStatic
         fun defaultCurrencyLocale(): Locale {
-            if (Purchases.isConfigured) {
-                return Purchases.sharedInstance.currencyLocaleForStorefrontCountryCode()
+            val storefrontCountryCode = if (Purchases.isConfigured) {
+                Purchases.sharedInstance.storefrontCountryCode
+            } else {
+                null
             }
-            return Locale.getDefault()
+            return CurrencyLocaleResolver.resolve(
+                storefrontCountryCode = storefrontCountryCode,
+                locale = Locale.getDefault(),
+            )
         }
     }
 }
