+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/data/MockPurchasesType.kt
@@ -11,7 +11,6 @@ import com.revenuecat.purchases.customercenter.CustomerCenterConfigData
 import com.revenuecat.purchases.customercenter.CustomerCenterListener
 import com.revenuecat.purchases.models.StoreProduct
 import com.revenuecat.purchases.virtualcurrencies.VirtualCurrencies
-import java.util.Locale
 
 /**
  * Mock implementation of [PurchasesType] for previews and test data
@@ -54,10 +53,5 @@ internal class MockPurchasesType(
     }
     override fun syncPurchases() {
         // No-op for mock
-    }
-
-    override fun currencyLocaleForStorefrontCountryCode(storefrontCountryCode: String?, locale: Locale): Locale {
-        // Just return the current locale for mock
-        return locale
     }
 }
